profile saving logic changed

+++ src/main/java/com/spring/jwt/repository/UserRepository.java
@@ -10,7 +10,7 @@ import java.util.Map;
 import java.util.Optional;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, Long> {
+public interface UserRepository extends JpaRepository<User, Integer> {
 
     User findByEmail(String email);
 
@@ -19,5 +19,5 @@ public interface UserRepository extends JpaRepository<User, Long> {
     Optional<User> findByMobileNumber(@Param("mobileNumber") Long mobileNumber);
     
     @Query(value = "SELECT * FROM users WHERE user_id = :id", nativeQuery = true)
-    Map<String, Object> findRawUserById(@Param("id") Long id);
+    Map<String, Object> findRawUserById(@Param("id") Integer id);
 }
